requirements.txt
## Core
python-dotenv>=1.0.0
pydantic>=2.0.0

## Orchestration
langgraph>=0.2.0
langchain-core>=0.2.0

## Gemini AI
google-generativeai>=0.8.0

## HTTP / async
httpx>=0.27.0
aiofiles>=23.0.0

## PDF processing
pypdf>=4.0.0

## Streamlit UI
streamlit>=1.38.0
Pillow>=10.0.0
